+++ test/bin/init.ps1
@@ -33,8 +33,8 @@ $buildVariables = ( Get-ChildItem -Path 'Env:' ).Where( { $_.Name -match '^(?:BH
 $buildVariables += ( Get-Variable -Name 'CI_*' -Scope 'Script' )
 $details = $buildVariables |
     Where-Object -FilterScript { $_.Name -notmatch 'EMAIL' } |
-        Sort-Object -Property 'Name' |
-            Format-Table -AutoSize -Property 'Name', 'Value' |
-                Out-String
+    Sort-Object -Property 'Name' |
+    Format-Table -AutoSize -Property 'Name', 'Value' |
+    Out-String
 Write-Host 'CI variables:'
 Write-Host $details -ForegroundColor DarkGray